+++ src/btrfs/replica.go
@@ -7,7 +7,7 @@ import (
 	"path"
 
 	"github.com/mitchellh/goamz/s3"
-	"github.com/pachyderm/pfs/lib/s3utils"
+	"github.com/pachyderm/pfs/src/s3utils"
 )
 
 type Pusher interface {
