+++ src/pipeline/pipeline.go
@@ -16,11 +16,11 @@ import (
 	"sync"
 
 	"github.com/fsouza/go-dockerclient"
-	"github.com/pachyderm/pfs/lib/btrfs"
-	"github.com/pachyderm/pfs/lib/concurrency"
-	"github.com/pachyderm/pfs/lib/container"
-	"github.com/pachyderm/pfs/lib/route"
-	"github.com/pachyderm/pfs/lib/s3utils"
+	"github.com/pachyderm/pfs/src/btrfs"
+	"github.com/pachyderm/pfs/src/concurrency"
+	"github.com/pachyderm/pfs/src/container"
+	"github.com/pachyderm/pfs/src/route"
+	"github.com/pachyderm/pfs/src/s3utils"
 )
 
 var (
