+++ src/shard/announce.go
